Add support for TLS/SSL mutual authentication
Add support in the TLS_FTPHandler to check a client certificate.  This
type of support strengthens the security between the client and the
server, only allowing clients with a valid certificate to connect to the
server.

Updated the api.rst file with the two new configurable options to make
client authentication work